+++ src/main/java/org/example/Constants.java
@@ -1,0 +1,8 @@
+package org.example;
+
+public class Constants {
+    public static final String SLASH = "/";
+    public static final String COMMA = ",";
+    public static final int NONE = 0;
+    public static final int SINGLE = 1;
+}
