+++ src/main/java/org/example/RecipeBook.java
@@ -1,0 +1,17 @@
+package org.example;
+
+import org.example.models.Library;
+
+import java.util.List;
+
+public class RecipeBook {
+    private final List<Library> libraries;
+
+    public RecipeBook(List<Library> libraries) {
+        this.libraries = libraries;
+    }
+
+    public List<Library> getLibraries() {
+        return libraries;
+    }
+}
