+++ src/main/java/org/example/crafters/BaseCrafter.java
@@ -1,0 +1,10 @@
+package org.example.crafters;
+
+import org.example.enums.Classification;
+
+public class BaseCrafter extends Crafter {
+    @Override
+    public Classification type() {
+        return super.type();
+    }
+}
