+++ src/main/java/org/example/enums/Classification.java
@@ -1,0 +1,9 @@
+package org.example.enums;
+
+public enum Classification {
+    FIRE,
+    FUEGO,
+    WATER,
+    AGUA,
+    ALL;
+}
