+++ src/main/java/org/example/enums/QueryEnum.java
@@ -1,0 +1,8 @@
+package org.example.enums;
+
+public enum QueryEnum {
+    ELEMENTS_FROM_ZERO,
+    ELEMENTS,
+    MISSING_ELEMENTS,
+    HOW_MANY_ELEMENTS;
+}
