+++ src/main/java/org/example/loaders/Loader.java
@@ -1,0 +1,5 @@
+package org.example.loaders;
+
+public interface Loader<T> {
+    void loadFile();
+}
