+++ src/main/java/org/example/models/Library.java
@@ -1,0 +1,4 @@
+package org.example.models;
+
+public record Library(String originTable, Recipe recipe) {
+}
