+++ src/main/java/org/example/models/QuantityElements.java
@@ -1,0 +1,9 @@
+package org.example.models;
+
+public class QuantityElements implements Queryable {
+    public final int number;
+
+    public QuantityElements(int number) {
+        this.number = number;
+    }
+}
