+++ src/main/java/org/example/models/Queryable.java
@@ -1,0 +1,4 @@
+package org.example.models;
+
+public interface Queryable {
+}
