+++ src/main/java/org/example/models/Recipe.java
@@ -1,0 +1,6 @@
+package org.example.models;
+
+import java.util.List;
+
+public record Recipe(Element give, float time, List<Element> ingredients) implements Queryable {
+}
