--- src/main/java/org/unlam/paradigmas/zeta/Constants.java
@@ -1,8 +1,0 @@
-package org.unlam.paradigmas.zeta;
-
-public class Constants {
-    public static final String SLASH = "/";
-    public static final String COMMA = ",";
-    public static final int NONE = 0;
-    public static final int SINGLE = 1;
-}
