--- src/main/java/org/unlam/paradigmas/zeta/crafters/BaseCrafter.java
@@ -1,10 +1,0 @@
-package org.unlam.paradigmas.zeta.crafters;
-
-import org.unlam.paradigmas.zeta.enums.Classification;
-
-public class BaseCrafter extends Crafter {
-    @Override
-    public Classification type() {
-        return super.type();
-    }
-}
