--- src/main/java/org/unlam/paradigmas/zeta/enums/Classification.java
@@ -1,9 +1,0 @@
-package org.unlam.paradigmas.zeta.enums;
-
-public enum Classification {
-    FIRE,
-    FUEGO,
-    WATER,
-    AGUA,
-    ALL;
-}
