--- src/main/java/org/unlam/paradigmas/zeta/enums/QueryEnum.java
@@ -1,8 +1,0 @@
-package org.unlam.paradigmas.zeta.enums;
-
-public enum QueryEnum {
-    ELEMENTS_FROM_ZERO,
-    ELEMENTS,
-    MISSING_ELEMENTS,
-    HOW_MANY_ELEMENTS;
-}
