--- src/main/java/org/unlam/paradigmas/zeta/models/Library.java
@@ -1,4 +1,0 @@
-package org.unlam.paradigmas.zeta.models;
-
-public record Library(String originTable, Recipe recipe) {
-}
