--- src/main/java/org/unlam/paradigmas/zeta/models/QuantityElements.java
@@ -1,9 +1,0 @@
-package org.unlam.paradigmas.zeta.models;
-
-public class QuantityElements implements Queryable {
-    public final int number;
-
-    public QuantityElements(int number) {
-        this.number = number;
-    }
-}
