--- src/main/java/org/unlam/paradigmas/zeta/models/Queryable.java
@@ -1,4 +1,0 @@
-package org.unlam.paradigmas.zeta.models;
-
-public interface Queryable {
-}
