--- src/main/java/org/unlam/paradigmas/zeta/models/Recipe.java
@@ -1,6 +1,0 @@
-package org.unlam.paradigmas.zeta.models;
-
-import java.util.List;
-
-public record Recipe(Element give, float time, List<Element> ingredients) implements Queryable {
-}
